Separated home and settings screen and added destination form to settings page
