+++ .expo-shared/assets.json
@@ -6,5 +6,5 @@
   "35b43002db9a83a41d6efba0d05109e32da0af1ec606ed427f6f4dae8a1148a6": true,
   "89ed26367cdb9b771858e026f2eb95bfdb90e5ae943e716575327ec325f39c44": true,
   "e816a2ca863b658024d1d72fad3c20bba301fbee6aff9668242d76ef5e459a72": true,
-  "fe5b9ddea615585d6b9eb6ffa7f97945363922e7d01c9cab04a4fbb13bd5b2ba": true
+  "1a96e4888cdd4f6a8ce81a41e0ce9891e8d5e03ff0913d843ab2c994ea9198b6": true
 }
