+++ package.json
@@ -22,6 +22,7 @@
     "test": "jest"
   },
   "dependencies": {
+    "@react-native-community/async-storage": "^1.5.0",
     "expo": "^33.0.0",
     "react": "16.8.3",
     "react-dom": "^16.8.6",
